Odd number of elements in hash list.
On Thu, 26 Mar 1998, Ed Jordan wrote:

> Perl should probably give a more specific warning where a hash is
> initialised with a single element which is a hash-ref, 

Okay. :-)

This patch changes the text of the warning given when warnings are on and
a hash is initialized with a single element which is either a hashref or
arrayref, typically like this.

    %hash = { qw/ I used curlies instead of parens / };
    %hash = [ qw/ This mistake is somewhat less common / ];

(At the same time, I've exorcized the oxymoron "hash list" from the
diagnostics.) 

It should be okay to add to the _05 release, since it's not a _new_
warning, just a change in the text of the existing one.

p5p-msgid: Pine.GSO.3.96.980328151929.29336D-100000@user2.teleport.com