typo-fix and suggestion for perlguts.pod
In the description of 'Stashes and Globs', perlguts suggests that
a GV contains references to both a File Handle and a Directory
Handle. I think it would be more accurate to replace these with a
single item such as 'I/O Handle', since that is what is stored in
the GP itself (ie at the same level as the other things mentioned).

Patch attached also fixes a minor typo. This is over perl5.004m5t1.


p5p-msgid: 199803051543.QAA03097@dorlas.elsevier.nl