Allow more lenient switch processing
This patch fixes two things:

	1. scripts beginning with #! that mention perl on the first line
	   and have a -i flag stop reading flags after the -i.  So,
		#!   perl -iorig -nl
	   won't see the -nl

	2. On AIX (shudder), scripts that start like #!/bin/perl -l -n -s
	   and are executed directly die with
		Unrecognized switch: - -n -s  (-h will show valid options).
	   This is because on AIX, what gets run is
		/bin/perl '-l -n -s' <script_name>

	   I don't think it does any harm to allow args like '-l -n -s', with
	   single blanks between flags to be valid, since presently they
	   aren't.  Two blanks between switches ends switch processing.
	   But perhaps it shouldn't.  Another thing that really bothers me
	   is the double switch processing that goes on with #! lines: first
	   when argv is scanned and again when the #! line is scanned.  It
	   really shouldn't happen, and can cause unexpected results.  Toke.c
	   should probably skip over the first part of the #! line that
	   exactly matches argv[1].  I think that was Chip's idea from a
	   while back.

The patch does _not_ fix the error msg that results from

	#!/bin/perl -z

	Can't emulate -z on #! line at /tmp/t line 1.

This msg should be "Invalid option -z on #! line ...", but I presume
other valid options can't be emulated on the #! line, hence this msg.
I'm a bit confused by the split option handling duties shared by the
reswitch code and the moreswitches sub in perl.c, so I'm not sure what
the right thing to do is.  It looks like moreswitches is mainly for
#! line flag processing, but does double duty for argv switch processing.
I considered making moreswitches take an additional arg that tells it
whether it's doing argv or #! processing, but became lazy: but in a
bad way :)

[ This morass of switch processing code should be rewritten, IMOSVHO.
  And much better commented. ]
 
p5p-msgid: 199803251638.LAA22664@gateway.grumman.com