bidirectional pipe warning blues
When Perl detects the use of a bidirectional pipe it will warn about it
(if warnings are enabled) and silently turn it into a writeable pipe.

For example once perl is finished with this bit of code 

    open (F, "| true |") ;

it will actually send "true |" to the shell for execution.

On my SunOS box this isn't a problem. Unfortunately Linux which uses
bash is another story.  It spits out this warning

    sh: -c: line 2: syntax error: unexpected end of file

Here is a patch to shut up the warning by zapping the terminating "|".

p5p-msgid: 9804082151.AA20399@claudius.bfsec.bt.co.uk