New Text::ParseWords
I've finally taken the time to fix several embarassing problems with
Text::ParseWords (most of them caused by yours truly).  This version
should be included in the Perl distribution as soon as humanly
possible.

Bugs fixed in this release include (but are not limited to):

	-- Handling of trailing null fields and fields
	   resolving to "false"

	-- Unpaired backslashes no longer cause syntax errors

	-- No Carp dependencies, does not die() on error

	-- No longer an autoloader, since use of the module
	   generally requires 80-90% of the module be loaded anyway

Functional improvements/changes include

	-- new @nested_quotewords() function which returns list
	   of lists, rather than a single long list

	-- parsing is no longer a character-by-character affair,
	   so performance is significantly improved

	-- semantics of $keep are expanded to include /bin/sh-like
	   collapsing of backslashes and special "delimiters" keyword
	   for preserving delimiters in addition to tokens

	-- additional documentation included, documentation updated

	-- many more regression tests in parsewords.t

It's better in all ways and I'm tired of fielding bug reports on the old
version.  Let's use this one!