perl compile fix for AIX 4.3
Perl does not compile under AIX 4.3 due to two problems. First,
in an attempt to make the DNS resolver thread safe h_errno is
not a simple variable any more, it is a define. This causes the
definition in pp_sys.c to fail. Second the XCOFF header files
remove some useful definitions that the dynamic loader in dl_aix.xs
did rely upon. The following is a patch that fixes both problems.

p5p-msgid: 199804261611.SAA34728@ans.helios.de