replace all "define\t" with "define " except cpan/, dist/ and Configure
'#define' has seven characters, so following it with a \t makes it look
like '#define ' when it is not, which then frustrates attempts to find
where a given define is. If you *know* then you do a

    git grep -P 'define\s+WHATEVER'

but if don't or you forget, you can get very confused trying to find
where a given define is located. This fixes all such cases so they
actually are 'define WHATEVER' instead.

View this patch with -w and you will see "no changes" except for
config_h.SH which needed to be changed to ensure that it produces output
that doesn't replicate the problem.

If this patch is getting in your way with blame analysis use the
-w option.

This patch does *not* include files in dist/ nor Configure related
files, those will be fixed in a subsequent patch.

This patch does *not* touch anything in cpan/ as we don't own that code.
Changes to files in dist/ will come in a subsequent patch.