dist/IO: replace "define\t" with "define ", sync versions, update ChangeLog
This cleans up the defines in dist/IO/poll.h.

At the same time I noticed that the version data in the different
modules in dist/IO varied from v1.48 to v1.49, which is a bit confusing.
The version for the dist comes from IO/IO.pm which was marked v1.49, but
this version was not mentioned in the Changelog. So I bumped the version
in all of the files to v1.50 and added a ChangeLog entry. The last
release on CPAN is v1.48 when all the versions where synchronized. At
least part of this version bump is to satisfy t/porting/cmp_version.t