regen/regcomp.pl - make robust to line continuations in header files
I noticed that tests fail if some of these defines are changed to use a
line continuation. This fixes the problem.

For instance, some of the defines that are parsed by this code are
rather long and if they get broken into a multiline statement with
C pre processor style line continuations they would get dropped from
regnodes.h and our debug output.